+++ requirements.txt
@@ -1,3 +1,3 @@
 kafka-python
 pyspark
-searchtweets
+tweepy
